Update from the failed branch where we try to replace core exec_dep to this one.
The main problem cannot have that replaced is due to hop estimation is
not exact. There are cases after chain decomposition, we simply cannot
find how many hops on the chain (because hops are longest distance, not
shortest). This doesn't seem easily solvable unless we start to record
distance within a chain.